vehicle-service/requirements.txt
flask==2.0.1
pyjwt==2.1.0
requests==2.26.0
